+++ crates/soundlog/src/lib.rs
@@ -6,4 +6,5 @@ pub mod vgm;
 
 pub use binutil::ParseError;
 pub use vgm::command::*;
+pub use vgm::stream::StreamResult as VgmStreamResult;
 pub use vgm::{VgmBuilder, VgmCallbackStream, VgmDocument, VgmExtraHeader, VgmHeader, VgmStream};